Break the netSdkTaskUtils <-> TaskHelper import cycle
`netSdkTaskUtils` imported `getDefaultContainerName` from `TaskHelper`, which
imports `NetSdkRunTaskProvider`, which imports `netSdkTaskUtils` back. The
provider module instantiates `new NetSdkRunTaskProvider()` at module scope, so
whenever `netSdkTaskUtils` was the module that entered the cycle first, the
constructor read `NetSdkRunTaskType` while it was still in the temporal dead
zone:

    ReferenceError: Cannot access 'NetSdkRunTaskType' before initialization

Nothing imported `netSdkTaskUtils` directly before, so the cycle was latent;
the new `netSdkTaskUtils.test.ts` is such an entry point, and whether it hits
the bad order depends on the platform's test-file glob ordering (it reproduced
on the Linux CI agents but not on Windows).

`getDefaultContainerName`/`getDefaultImageName` are pure name helpers over
`getValidImageName`, with no other `TaskHelper` dependencies, so move them to
`utils/getValidImageName.ts` (a leaf module) and update the importers. This
removes the cycle rather than papering over the ordering.

Copilot-Session: 2ff5afb3-accd-442d-a1cc-c8f995b7122d
